+++ test/test_async.js
@@ -298,8 +298,15 @@ test("callbacks are ordered correctly", function() {
       { name: 'clear', from: 'yellow', to: 'green'  },
     ],
     callbacks: {
-      onchangestate: function(event,from,to) { called.push('onchange from ' + from + ' to ' + to); },
 
+      // generic callbacks
+      onbeforeevent: function(event,from,to) { called.push('onbefore(' + event + ')'); },
+      onafterevent:  function(event,from,to) { called.push('onafter('  + event + ')'); },
+      onleavestate:  function(event,from,to) { called.push('onleave('  + from  + ')'); },
+      onenterstate:  function(event,from,to) { called.push('onenter('  + to    + ')'); },
+      onchangestate: function(event,from,to) { called.push('onchange(' + from + ',' + to + ')'); },
+
+      // specific state callbacks
       onentergreen:  function() { called.push('onentergreen');                             },
       onenteryellow: function() { called.push('onenteryellow');                            },
       onenterred:    function() { called.push('onenterred');                               },
@@ -307,6 +314,7 @@ test("callbacks are ordered correctly", function() {
       onleaveyellow: function() { called.push('onleaveyellow'); return StateMachine.ASYNC; },
       onleavered:    function() { called.push('onleavered');    return StateMachine.ASYNC; },
 
+      // specific event callbacks
       onbeforewarn:  function() { called.push('onbeforewarn');                },
       onbeforepanic: function() { called.push('onbeforepanic');               },
       onbeforecalm:  function() { called.push('onbeforecalm');                },
@@ -319,20 +327,20 @@ test("callbacks are ordered correctly", function() {
   });
 
   called = [];
-  fsm.warn();       deepEqual(called, ['onbeforewarn', 'onleavegreen']);
-  fsm.transition(); deepEqual(called, ['onbeforewarn', 'onleavegreen', 'onenteryellow', 'onchange from green to yellow', 'onafterwarn']);
+  fsm.warn();       deepEqual(called, ['onbeforewarn', 'onbefore(warn)', 'onleavegreen', 'onleave(green)']);
+  fsm.transition(); deepEqual(called, ['onbeforewarn', 'onbefore(warn)', 'onleavegreen', 'onleave(green)', 'onenteryellow', 'onenter(yellow)', 'onchange(green,yellow)', 'onafterwarn', 'onafter(warn)']);
 
   called = [];
-  fsm.panic();      deepEqual(called, ['onbeforepanic', 'onleaveyellow']);
-  fsm.transition(); deepEqual(called, ['onbeforepanic', 'onleaveyellow', 'onenterred', 'onchange from yellow to red', 'onafterpanic']);
+  fsm.panic();      deepEqual(called, ['onbeforepanic', 'onbefore(panic)', 'onleaveyellow', 'onleave(yellow)']);
+  fsm.transition(); deepEqual(called, ['onbeforepanic', 'onbefore(panic)', 'onleaveyellow', 'onleave(yellow)', 'onenterred', 'onenter(red)', 'onchange(yellow,red)', 'onafterpanic', 'onafter(panic)']);
 
   called = [];
-  fsm.calm();       deepEqual(called, ['onbeforecalm', 'onleavered']);
-  fsm.transition(); deepEqual(called, ['onbeforecalm', 'onleavered', 'onenteryellow', 'onchange from red to yellow', 'onaftercalm']);
+  fsm.calm();       deepEqual(called, ['onbeforecalm', 'onbefore(calm)', 'onleavered', 'onleave(red)']);
+  fsm.transition(); deepEqual(called, ['onbeforecalm', 'onbefore(calm)', 'onleavered', 'onleave(red)', 'onenteryellow', 'onenter(yellow)', 'onchange(red,yellow)', 'onaftercalm', 'onafter(calm)']);
 
   called = [];
-  fsm.clear();      deepEqual(called, ['onbeforeclear', 'onleaveyellow']);
-  fsm.transition(); deepEqual(called, ['onbeforeclear', 'onleaveyellow', 'onentergreen', 'onchange from yellow to green', 'onafterclear']);
+  fsm.clear();      deepEqual(called, ['onbeforeclear', 'onbefore(clear)', 'onleaveyellow', 'onleave(yellow)']);
+  fsm.transition(); deepEqual(called, ['onbeforeclear', 'onbefore(clear)', 'onleaveyellow', 'onleave(yellow)', 'onentergreen', 'onenter(green)', 'onchange(yellow,green)', 'onafterclear', 'onafter(clear)']);
 
 });