+++ test/test_initialize.js
@@ -4,21 +4,25 @@ module("special initialization options", {
 
   setup: function() {
     this.called = [];
-    this.onchangestate   = function(event,from,to) { this.called.push('onchange from ' + from + ' to ' + to); };
-    this.onbeforeinit    = function()              { this.called.push("onbeforeinit");                        };
-    this.onafterinit     = function()              { this.called.push("onafterinit");                         };
-    this.onbeforestartup = function()              { this.called.push("onbeforestartup");                     };
-    this.onafterstartup  = function()              { this.called.push("onafterstartup");                      };
-    this.onbeforepanic   = function()              { this.called.push("onbeforepanic");                       };
-    this.onafterpanic    = function()              { this.called.push("onafterpanic");                        };
-    this.onbeforecalm    = function()              { this.called.push("onbeforecalm");                        };
-    this.onaftercalm     = function()              { this.called.push("onaftercalm");                         };
-    this.onenternone     = function()              { this.called.push("onenternone");                         };
-    this.onentergreen    = function()              { this.called.push("onentergreen");                        };
-    this.onenterred      = function()              { this.called.push("onenterred");                          };
-    this.onleavenone     = function()              { this.called.push("onleavenone");                         };
-    this.onleavegreen    = function()              { this.called.push("onleavegreen");                        };
-    this.onleavered      = function()              { this.called.push("onleavered");                          };
+    this.onbeforeevent   = function(event,from,to) { this.called.push('onbefore(' + event + ')');           },
+    this.onafterevent    = function(event,from,to) { this.called.push('onafter('  + event + ')');           },
+    this.onleavestate    = function(event,from,to) { this.called.push('onleave('  + from  + ')');           },
+    this.onenterstate    = function(event,from,to) { this.called.push('onenter('  + to    + ')');           },
+    this.onchangestate   = function(event,from,to) { this.called.push('onchange(' + from + ',' + to + ')'); };
+    this.onbeforeinit    = function()              { this.called.push("onbeforeinit");                      };
+    this.onafterinit     = function()              { this.called.push("onafterinit");                       };
+    this.onbeforestartup = function()              { this.called.push("onbeforestartup");                   };
+    this.onafterstartup  = function()              { this.called.push("onafterstartup");                    };
+    this.onbeforepanic   = function()              { this.called.push("onbeforepanic");                     };
+    this.onafterpanic    = function()              { this.called.push("onafterpanic");                      };
+    this.onbeforecalm    = function()              { this.called.push("onbeforecalm");                      };
+    this.onaftercalm     = function()              { this.called.push("onaftercalm");                       };
+    this.onenternone     = function()              { this.called.push("onenternone");                       };
+    this.onentergreen    = function()              { this.called.push("onentergreen");                      };
+    this.onenterred      = function()              { this.called.push("onenterred");                        };
+    this.onleavenone     = function()              { this.called.push("onleavenone");                       };
+    this.onleavegreen    = function()              { this.called.push("onleavegreen");                      };
+    this.onleavered      = function()              { this.called.push("onleavered");                        };
   }
 
 });
@@ -47,7 +51,17 @@ test("initial state can be specified", function() {
       { name: 'calm',  from: 'red',   to: 'green' }
   ]});
   equal(this.current, 'green');
-  deepEqual(this.called, ["onbeforestartup", "onleavenone", "onentergreen", "onchange from none to green", "onafterstartup"]);
+  deepEqual(this.called, [
+    "onbeforestartup",
+    "onbefore(startup)",
+    "onleavenone",
+    "onleave(none)",
+    "onentergreen",
+    "onenter(green)",
+    "onchange(none,green)",
+    "onafterstartup",
+    "onafter(startup)"
+  ]);
 });
 
 //-----------------------------------------------------------------------------
@@ -61,7 +75,17 @@ test("startup event name can be specified", function() {
       { name: 'calm',  from: 'red',   to: 'green' }
   ]});
   equal(this.current, 'green');
-  deepEqual(this.called, ["onbeforeinit", "onleavenone", "onentergreen", "onchange from none to green", "onafterinit"]);
+  deepEqual(this.called, [
+    "onbeforeinit",
+    "onbefore(init)",
+    "onleavenone",
+    "onleave(none)",
+    "onentergreen",
+    "onenter(green)",
+    "onchange(none,green)",
+    "onafterinit",
+    "onafter(init)"
+  ]);
 });
 
 //-----------------------------------------------------------------------------
@@ -80,7 +104,17 @@ test("startup event can be deferred", function() {
   this.init();
 
   equal(this.current, 'green');
-  deepEqual(this.called, ["onbeforeinit", "onleavenone", "onentergreen", "onchange from none to green", "onafterinit"]);
+  deepEqual(this.called, [
+    "onbeforeinit",
+    "onbefore(init)",
+    "onleavenone",
+    "onleave(none)",
+    "onentergreen",
+    "onenter(green)",
+    "onchange(none,green)",
+    "onafterinit",
+    "onafter(init)"
+  ]);
 });
 
 //-----------------------------------------------------------------------------